WPG support, part 1, Sw: Move Textbox link to UNO property
Before TextBoxes (Shape + Frame) were linked in the SwFrameFormat
class. I had to move this link, because group shapes have only
one framefomat for the whole group and the child shapes have not
got own fomat, so this can not make that possible to each shape
in the group have an associated textbox like in OOXML. After this
patch the shape has a property namely TextBoxContent, which allows
to reach the associated textframe via UNO. This was neccessary for
filters as well, because the old way (casting the XShape to XRange)
could not work with group shapes, because the above mentioned
problem. In addition the SwFormatContent which provided that cast,
could not be used for the group shapes because one format means
one content too. So that is why that modification was neccesary.

Please note this commit is a part of a commit range, and this one
is the fist part so follow ups will come. Here is an overview:

Part 1: Move textbox link to UNO <-This commit
Part 2: Make Writerfiler and OOX capable to handle WPG,
Part 3: Extend ODF standard and filter if needed.
Part 4: WW8 Export implementation of WPG support.

Note: Unit test for this part is under develeopment, WIP patch!

+++ sw/source/core/unocore/unodraw.cxx
@@ -1163,13 +1163,20 @@ void SwXShape::setPropertyValue(const OUString& rPropertyName, const uno::Any& a
             }
             else if (pEntry->nWID == FN_TEXT_BOX)
             {
-                bool bValue(false);
-                aValue >>= bValue;
-                if (bValue)
-                    SwTextBoxHelper::create(pFormat);
-                else
-                    SwTextBoxHelper::destroy(pFormat);
+                if (pEntry->nMemberId == MID_TEXTBOX)
+                {
+                    bool bValue(false);
+                    aValue >>= bValue;
+                    if (bValue)
+                        SwTextBoxHelper::createTextBox(mxShape, pFormat->GetDoc());
+                    else
+                        SwTextBoxHelper::destroy(pFormat);
+                }
 
+                if (pEntry->nMemberId == MID_TEXTBOX_CONTENT)
+                {
+                    m_xTextBox.set(aValue.get<uno::Reference<css::text::XTextFrame>>());
+                }
             }
             else if (pEntry->nWID == RES_CHAIN)
             {
@@ -1516,8 +1523,15 @@ uno::Any SwXShape::getPropertyValue(const OUString& rPropertyName)
                 }
                 else if (pEntry->nWID == FN_TEXT_BOX)
                 {
-                    bool bValue = SwTextBoxHelper::isTextBox(pFormat, RES_DRAWFRMFMT);
-                    aRet <<= bValue;
+                    if (pEntry->nMemberId == MID_TEXTBOX)
+                    {
+                        bool bValue = SwTextBoxHelper::isTextBox(mxShape);
+                        aRet <<= bValue;
+                    }
+                    else if (pEntry->nMemberId == MID_TEXTBOX_CONTENT)
+                    {
+                        aRet = uno::Any(m_xTextBox);
+                    }
                 }
                 else if (pEntry->nWID == RES_CHAIN)
                 {
@@ -1795,7 +1809,7 @@ uno::Sequence< beans::PropertyState > SwXShape::getPropertyStates(
                      (pEntry->nMemberId == MID_FRMSIZE_REL_HEIGHT_RELATION ||
                       pEntry->nMemberId == MID_FRMSIZE_REL_WIDTH_RELATION))
                 pRet[nProperty] = beans::PropertyState_DIRECT_VALUE;
-            else if (pEntry->nWID == FN_TEXT_BOX)
+            else if (pEntry->nWID == FN_TEXT_BOX && pEntry->nMemberId == MID_TEXTBOX)
             {
                 // The TextBox property is set, if we can find a textbox for this shape.
                 if (pFormat && SwTextBoxHelper::isTextBox(pFormat, RES_DRAWFRMFMT))
